rename azure_services to azure_vm_services
