+++ mlos_bench/mlos_bench/services/remote/azure/__init__.py
@@ -7,7 +7,7 @@ Azure-specific benchmark environments for mlos_bench.
 """
 
 from mlos_bench.services.remote.azure.azure_auth import AzureAuthService
-from mlos_bench.services.remote.azure.azure_services import AzureVMService
+from mlos_bench.services.remote.azure.azure_vm_services import AzureVMService
 from mlos_bench.services.remote.azure.azure_fileshare import AzureFileShareService
 from mlos_bench.services.remote.azure.azure_saas import AzureSaaSConfigService
